Moved withdraw liquidity to its own widget
